+++ packages/components/src/index.tsx
@@ -28,6 +28,7 @@ export * from './tag'
 export * from './text'
 export * from './toast'
 export * from './tokens'
+export * from './tooltip'
 export * from './user-list'
 // eslint-disable-next-line simple-import-sort/exports
 export { config } from './tamagui.config'