+++ packages/status-js/CHANGELOG.md
@@ -1,5 +1,72 @@
 # @status-im/js
 
+## 0.6.6
+
+### Patch Changes
+
+- ef5c31f: trim whitespace of encoded emoji
+
+## 0.6.5
+
+### Patch Changes
+
+- a7008b5: handle segmented waku messages
+- 4106248: optimize handling of token owner waku messages
+
+## 0.6.4
+
+### Patch Changes
+
+- 8249e9f: fix community schema in encode-url-data.ts
+
+## 0.6.3
+
+### Patch Changes
+
+- 13f7ed5: Update encode-url-data.ts
+- 18504de: fix user schema in `encode-url-data.ts`
+
+## 0.6.2
+
+### Patch Changes
+
+- f9af5e0: validate encoded data
+
+## 0.6.1
+
+### Patch Changes
+
+- e863523: update `RequestClient` options
+
+## 0.6.0
+
+### Minor Changes
+
+- 195e3f9: validate metadata for communities with owner token
+
+### Patch Changes
+
+- 7da6546: iterate metadata sequentially
+
+## 0.5.1
+
+### Patch Changes
+
+- 6b4e6b1: stop fetching early
+
+## 0.5.0
+
+### Minor Changes
+
+- 9a495fd: handle updates for encrypted communities
+- 02abf34: update `protos`
+
+## 0.4.6
+
+### Patch Changes
+
+- 22e727e: set MAX_DATA_LENGTH
+
 ## 0.4.5
 
 ### Patch Changes